python_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
PyJWT==2.10.1
psutil==6.1.0
